Merged in TPI-88-import-level-geometry (pull request #28)
Level Geo Import

* Level Geo Import

Imported (and cleaned up) the asset package. Package contained some of my testing scripts, but those and all references to them have been scrubbed.

Began swapping/placing a few assets in the scene.

When reviewing, you will NOT see the leve geometry in the scene, you must locate the prefab and view it in isolation, or drag the prefab into the scene. The geo can be added to the scene duing a future ticket when we know there will not be any other scene level changes and thus scene conflicts.


Approved-by: Carolina Naoum Junqueira