Merged in TPI-285-adjust-walls-and-populate (pull request #132)
TPI-285 adjust walls and populate

* Making the Space a Space
Adjusted the wall thickness of all interior walls ro better accomodate doors and frames
Added initial materials to the scene to give a better idea on the visual colour language of the final
Placed the wooden shelving and dyed it black so that it doesn't stick out as an earth tone next to these loud colours
Ran trim thoughout the space
Created initial lighting to give a general sense of mood that I'm aiming for
Replaced all placeholder volumes except for: lobby table, front door, and shelving in back room
Moved the Piercing Artist as the shelving no longer accomodates him on that side. This possition does not need to be permanent
Added ceiilings throughout the space!
Create a massive light bocking voluume around the environment to eliminate light bleed
Adjusted URP blance to alow for a few more realtime lights
Imported and created varions for wall paint materials
Create prefabs for the striaght and corner trim objects

Began experimenting with reflection probes. These might not stick a…
* Haha, unbroke the Piercing door

* Created a lighting rig for the object viewer, since there is no longer a directional light
Deleted duplicate the procedure tray, which was allowing you to open an inspection but the object would be invisible (because the dupe was static).


Approved-by: Juhwan Seo
Approved-by: Karandeep Sandhu
Approved-by: Salman Nouman Abulqasim