Merged in TPI-398-asset-import-23jun2025 (pull request #203)
TPI-398 asset import 23jun2025

* TPI- 398 asset import

* TPI-325 - added the skate sign to the reception desk

* TPI-260 imported the enzymax spray bottle

* added to the toolchest in the tattoo and piercing areas

* TPI-311 imported light fixture model

* TPI-344 imported the decorative table

* added the decorative table to the lobby

* TPI-396 imported snake

* place snake on the reception desk

* TPI-397 imported folding table and chair and fridge

* added the table, chairs and fridge to the reprocessing area

* Light Models Config
Created some assorted light prefabs. Added these prefabs to the bathroom, the hall, the lobby, the piercing room, and hte tattoo room.
Created a material for the light to be more easily adjustable if the lighting changes
Adjusted the WoodenFloorShelves prefab to have a single top shelf and closed doors, which eliminates the need to populate all the shelves while still implying storage


Approved-by: Karandeep Sandhu